@@ -8,6 +8,8 @@
     "watch": "cross-env tsc --watch",
     "test": "cross-env jest",
     "coverage": "npm test -- --coverage",
+    "lint": "./node_modules/eslint/bin/eslint.js . --ext .js",
+    "lint:fix": "npm run lint -- --fix",
     "start": "cross-env npm run build && medusa start",
     "start:custom": "cross-env npm run build && node --preserve-symlinks --trace-warnings index.js",
     "dev": "cross-env npm run build:server && medusa develop"
@@ -22,6 +24,7 @@
     "cors": "^2.8.5",
     "dotenv": "16.3.1",
     "express": "^4.17.2",
+    "medusa-core-utils": "^1.2.1",
     "medusa-interfaces": "^1.3.8",
     "typeorm": "^0.3.16"
   },
@@ -35,8 +38,12 @@
     "@types/node": "^17.0.8",
     "babel-preset-medusa-package": "^1.1.19",
     "cross-env": "^7.0.3",
-    "eslint": "^6.8.0",
+    "eslint": "^8.37.0",
+    "eslint-config-prettier": "^8.8.0",
+    "eslint-plugin-import": "^2.27.5",
+    "eslint-plugin-prettier": "^4.2.1",
     "jest": "^27.3.1",
+    "prettier": "^2.8.7",
     "rimraf": "^3.0.2",
     "ts-jest": "^27.0.7",
     "ts-loader": "^9.2.6",
